tea_project/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
idna==3.4
numpy==1.23.4
oauthlib==3.2.2
pandas==1.5.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
